fix for __init__.py
